finish fixing tests
