+++ app/frontend/src/features/auth/useAuthStore.test.ts
@@ -5,10 +5,10 @@ import { act } from "react-test-renderer";
 import { service, SignupArguments } from "../../service";
 import { addDefaultUser } from "../../test/utils";
 import useAuthStore, { usePersistedState } from "./useAuthStore";
+import wrapper from "../../test/hookWrapper";
 
 const getUserMock = service.user.getUser as jest.Mock;
 const getCurrentUserMock = service.user.getCurrentUser as jest.Mock;
-const updateProfileMock = service.user.updateProfile as jest.Mock;
 const passwordLoginMock = service.user.passwordLogin as jest.Mock;
 const tokenLoginMock = service.user.tokenLogin as jest.Mock;
 const signupMock = service.user.completeSignup as jest.Mock;
@@ -38,7 +38,7 @@ describe("usePersistedState hook", () => {
 
 describe("useAuthStore hook", () => {
   it("sets and clears an error", async () => {
-    const { result } = renderHook(() => useAuthStore());
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     act(() => result.current.authActions.authError("error1"));
     expect(result.current.authState.error).toBe("error1");
     act(() => result.current.authActions.clearError());
@@ -48,7 +48,7 @@ describe("useAuthStore hook", () => {
   it("logs out", async () => {
     logoutMock.mockResolvedValue(new Empty());
     addDefaultUser();
-    const { result } = renderHook(() => useAuthStore());
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     expect(result.current.authState.authenticated).toBe(true);
     await act(() => result.current.authActions.logout());
     expect(result.current.authState.authenticated).toBe(false);
@@ -60,8 +60,10 @@ describe("useAuthStore hook", () => {
 describe("passwordLogin action", () => {
   it("sets authenticated correctly", async () => {
     passwordLoginMock.mockResolvedValue({ jailed: false });
-    getUserMock.mockResolvedValue(null);
-    const { result } = renderHook(() => useAuthStore());
+    getUserMock.mockResolvedValue(defaultUser);
+    const { result } = renderHook(() => useAuthStore(), {
+      wrapper,
+    });
     expect(result.current.authState.authenticated).toBe(false);
     await act(() =>
       result.current.authActions.passwordLogin({
@@ -76,7 +78,7 @@ describe("passwordLogin action", () => {
       code: StatusCode.PERMISSION_DENIED,
       message: "Invalid username or password.",
     });
-    const { result } = renderHook(() => useAuthStore());
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     expect(result.current.authState.authenticated).toBe(false);
     await act(() =>
       result.current.authActions.passwordLogin({
@@ -94,8 +96,8 @@ describe("passwordLogin action", () => {
 describe("tokenLogin action", () => {
   it("sets authenticated correctly", async () => {
     tokenLoginMock.mockResolvedValue({ jailed: false });
-    getCurrentUserMock.mockResolvedValue(null);
-    const { result } = renderHook(() => useAuthStore());
+    getCurrentUserMock.mockResolvedValue(defaultUser);
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     expect(result.current.authState.authenticated).toBe(false);
     await act(() => result.current.authActions.tokenLogin("token"));
     expect(result.current.authState.authenticated).toBe(true);
@@ -105,7 +107,7 @@ describe("tokenLogin action", () => {
       code: StatusCode.PERMISSION_DENIED,
       message: "Invalid token.",
     });
-    const { result } = renderHook(() => useAuthStore());
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     expect(result.current.authState.authenticated).toBe(false);
     await act(() => result.current.authActions.tokenLogin("token"));
     expect(result.current.authState.authenticated).toBe(false);
@@ -118,8 +120,8 @@ describe("signup action", () => {
   //and getCurrentUser
   it("sets authenticated correctly", async () => {
     signupMock.mockResolvedValue({ jailed: false });
-    getCurrentUserMock.mockResolvedValue(null);
-    const { result } = renderHook(() => useAuthStore());
+    getCurrentUserMock.mockResolvedValue(defaultUser);
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     expect(result.current.authState.authenticated).toBe(false);
     await act(() => result.current.authActions.signup({} as SignupArguments));
     expect(result.current.authState.authenticated).toBe(true);
@@ -129,7 +131,7 @@ describe("signup action", () => {
       code: StatusCode.PERMISSION_DENIED,
       message: "Invalid token.",
     });
-    const { result } = renderHook(() => useAuthStore());
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     expect(result.current.authState.authenticated).toBe(false);
     await act(() => result.current.authActions.signup({} as SignupArguments));
     expect(result.current.authState.authenticated).toBe(false);
@@ -141,7 +143,7 @@ describe("updateJailStatus action", () => {
   it("sets jailed to true for jailed user", async () => {
     getIsJailedMock.mockResolvedValue({ isJailed: true });
     addDefaultUser();
-    const { result } = renderHook(() => useAuthStore());
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     await act(() => result.current.authActions.updateJailStatus());
     expect(result.current.authState.jailed).toBe(true);
     expect(result.current.authState.authenticated).toBe(true);
@@ -150,116 +152,9 @@ describe("updateJailStatus action", () => {
     getIsJailedMock.mockResolvedValue({ isJailed: false });
     getCurrentUserMock.mockResolvedValue({});
     addDefaultUser();
-    const { result } = renderHook(() => useAuthStore());
+    const { result } = renderHook(() => useAuthStore(), { wrapper });
     await act(() => result.current.authActions.updateJailStatus());
     expect(result.current.authState.jailed).toBe(false);
     expect(result.current.authState.authenticated).toBe(true);
-  });
-});
-
-describe("updateUserProfile action", () => {
-  it("updates the store with the latest user profile info", async () => {
-    addDefaultUser();
-    const {
-      aboutMe,
-      aboutPlace,
-      radius,
-      countriesLivedList,
-      countriesVisitedList,
-      gender,
-      name,
-      occupation,
-    } = defaultUser;
-    const newUserProfileData = {
-      // Unchanged data
-      name,
-      gender,
-      occupation,
-      aboutMe,
-      aboutPlace,
-      countriesLived: countriesLivedList,
-      // Changed data
-      city: "New York",
-      lat: 40.7306,
-      lng: -73.9352,
-      radius,
-      hostingStatus: 3,
-      languages: ["English", "Finnish", "Spanish"],
-      countriesVisited: [...countriesVisitedList, "United States"],
-    };
-    updateProfileMock.mockResolvedValue(new Empty());
-    getUserMock.mockResolvedValue({
-      ...newUserProfileData,
-      countriesLivedList,
-      countriesVisitedList: newUserProfileData.countriesVisited,
-    });
-    const { result } = renderHook(() => useAuthStore());
-
-    await act(() =>
-      result.current.profileActions.updateUserProfile(newUserProfileData)
-    );
-
-    expect(updateProfileMock).toHaveBeenCalledTimes(1);
-    expect(updateProfileMock).toHaveBeenCalledWith(newUserProfileData);
-    expect(getUserMock).toHaveBeenCalledTimes(1);
-    expect(getUserMock).toHaveBeenCalledWith("aapeli");
-
-    const state = result.current.authState;
-
-    // Things that have been updated are being reflected
-    expect(state.user).toMatchObject({
-      city: "New York",
-      hostingStatus: 3,
-      languages: ["English", "Finnish", "Spanish"],
-      countriesLivedList: ["Australia", "Finland", "Sweden", "United States"],
-      countriesVisitedList: ["Australia", "United States"],
-      lat: 40.7306,
-      lng: -73.9352,
-    });
-    // Things haven't been updated should remain the same
-    expect(state.user).toMatchObject({
-      name: "Aapeli Vuorinen",
-      gender: "Male",
-      occupation: "Mathematician",
-      aboutMe: "Some generic stuff.",
-      aboutPlace: "About Aapeli's place",
-      radius: 200,
-    });
-  });
-
-  it("does not update the existing user if the API call failed", async () => {
-    addDefaultUser();
-    updateProfileMock.mockRejectedValue(new Error("API error"));
-    const { result } = renderHook(() => useAuthStore());
-
-    expect(
-      act(() =>
-        result.current.profileActions.updateUserProfile({
-          ...defaultUser,
-          languages: defaultUser.languagesList,
-          countriesLived: ["Ecuador"],
-          countriesVisited: defaultUser.countriesVisitedList,
-        })
-      )
-    ).rejects.toThrow();
-
-    expect(result.current.authState.user).toEqual(defaultUser);
-  });
-
-  it("does not try to update profile if a user does not exist in store", async () => {
-    const { result } = renderHook(() => useAuthStore());
-    expect(
-      act(() =>
-        result.current.profileActions.updateUserProfile({
-          ...defaultUser,
-          languages: defaultUser.languagesList,
-          countriesLived: defaultUser.countriesLivedList,
-          countriesVisited: defaultUser.countriesVisitedList,
-        })
-      )
-    ).rejects.toBeDefined();
-
-    expect(updateProfileMock).not.toHaveBeenCalled();
-    expect(getUserMock).not.toHaveBeenCalled();
   });
 });
