+++ app/frontend/src/features/connections/friends/useFriendList.test.tsx
@@ -5,21 +5,10 @@ import { service } from "../../../service";
 import useFriendList from "./useFriendList";
 import { wait } from "../../../test/utils";
 import { getUser, listFriends } from "../../../test/serviceMockDefaults";
+import wrapper from "../../../test/hookWrapper";
 
 const listFriendsMock = service.api.listFriends as jest.Mock;
 const getUserMock = service.user.getUser as jest.Mock;
-
-const wrapper = ({ children }: { children: React.ReactNode }) => {
-  const client = new QueryClient({
-    defaultOptions: {
-      queries: {
-        retry: false,
-      },
-    },
-  });
-
-  return <QueryClientProvider client={client}>{children}</QueryClientProvider>;
-};
 
 beforeAll(() => {
   // Mock out console.error so the test output is less noisy when
