+++ app/frontend/src/test/hookWrapper.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+import { QueryClient, QueryClientProvider } from "react-query";
+import { MemoryRouter } from "react-router-dom";
+import AuthProvider from "../features/auth/AuthProvider";
+
+export default function hookWrapper({ children }: { children: any }) {
+  const client = new QueryClient({
+    defaultOptions: {
+      queries: {
+        retry: false,
+      },
+    },
+  });
+  return (
+    <MemoryRouter>
+      <QueryClientProvider client={client}>
+        <AuthProvider>{children}</AuthProvider>
+      </QueryClientProvider>
+    </MemoryRouter>
+  );
+}
